@@ -1,4 +1,4 @@
-import 'package:aqueduct/aqueduct.dart';
+import 'package:conduit/conduit.dart';
 import 'package:aqueduct_doc_to_obj/aqueduct_doc_to_obj.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'dart:convert';
@@ -6,9 +6,9 @@ part 'main.g.dart';
 
 @JsonSerializable()
 class Param {
-  final String name;
-  final String value;
-  final DateTime updateDate;
+  final String? name;
+  final String? value;
+  final DateTime? updateDate;
   Param({this.name, this.value, this.updateDate});
   factory Param.fromJson(Map<String, dynamic> json) => _$ParamFromJson(json);
   Map<String, dynamic> toJson() => _$ParamToJson(this);
@@ -16,22 +16,22 @@ class Param {
 
 class Person extends _Person {
   @DocToObj()
-  List<Param> _paramObj;
+  List<Param>? _paramObj;
 }
 
 class _Person {
   @Column(primaryKey: true)
-  String id;
+  String? id;
   @Column(defaultValue: "''")
-  String name;
-  DateTime birthDate;
-  Document _param;
+  String? name;
+  DateTime? birthDate;
+  Document? _param;
 }
 
 void main() {
   final e = Person();
   e.param = [Param(name: "test", value: "test", updateDate: DateTime.now())];
 
-  print('internal data \n${json.encode(e._param.data)}');
-  print('Person field data: \n Name ${e.param.first.name}');
+  print('internal data \n${json.encode(e._param?.data)}');
+  print('Person field data: \n Name ${e.param?.first.name}');
 }
