@@ -7,39 +7,47 @@ part of 'main.dart';
 // **************************************************************************
 
 extension PersonDocToParam on Person {
-  List<Param> get param {
+  List<Param>? get param {
     if (_paramObj == null) {
-      if (_param == null) {
-        return null;
-      } else {
-        final l = _param.data as List;
+      final docField = _param;
+      if (docField != null) {
+        final l = docField.data as List;
         return _paramObj =
             l.map((e) => Param.fromJson(e as Map<String, dynamic>)).toList();
+      } else {
+        return null;
       }
     } else {
       return _paramObj;
     }
   }
 
-  set param(List<Param> s) {
-    final l = s.map((e) => e.toJson()).toList();
-    _paramObj = s;
-    _param = Document(l);
+  set param(List<Param>? s) {
+    if (s != null) {
+      final l = s.map((e) => e.toJson()).toList();
+      _paramObj = s;
+      _param = Document(l);
+    } else {
+      _paramObj = null;
+      _param = null;
+    }
   }
 
   void readParamFromMap(Map<String, dynamic> object) {
-    if (object.containsKey('param')) {
+    if (object.containsKey('param') && object['param'] != null) {
       var l = object['param'] as List;
-      _paramObj =
+      final obj =
           l.map((e) => Param.fromJson(e as Map<String, dynamic>)).toList();
-      l = _paramObj.map((e) => e.toJson()).toList();
+      l = obj.map((e) => e.toJson()).toList();
       _param = Document(l);
+      _paramObj = obj;
     }
   }
 
   void paramToMap(Map<String, dynamic> map) {
-    if (_param != null) {
-      map['param'] = _param.data;
+    final docField = _param;
+    if (docField != null) {
+      map['param'] = docField.data;
     }
   }
 }
@@ -50,8 +58,8 @@ extension PersonDocToParam on Person {
 
 Param _$ParamFromJson(Map<String, dynamic> json) {
   return Param(
-    name: json['name'] as String,
-    value: json['value'] as String,
+    name: json['name'] as String?,
+    value: json['value'] as String?,
     updateDate: json['updateDate'] == null
         ? null
         : DateTime.parse(json['updateDate'] as String),
